Rebased with 703e17fa5cb4cb797ce17e8b1b564feb9d7f3751

+++ src/x32/assembler-x32.h
@@ -358,8 +358,8 @@ class Immediate BASE_EMBEDDED {
  public:
   explicit Immediate(int32_t value) : value_(value) {}
   explicit Immediate(Smi* value) {
-    ASSERT(SmiValuesAre31Bits());   // Only available for 31-bit SMI.
-    value_ = reinterpret_cast<intptr_t>(value);
+    ASSERT(SmiValuesAre31Bits());  // Only available for 31-bit SMI.
+    value_ = static_cast<int32_t>(reinterpret_cast<intptr_t>(value));
   }
 
  private:
